requirements.txt
prefect
prefect-cloud
prefect-dbt
prefect-gcp
dbt-core==1.10.2
dbt-bigquery==1.10.2
google-cloud-bigquery
google-cloud-storage
python-dotenv